requirements.txt
requests
pypandoc
pdf2image
Pillow
beautifulsoup4
rapidfuzz
sacrebleu
mistune
bert-score
